* commit '35da149d1c21ba4b5a24f71c9f1f98d45df1c7b0':
  Various fixes to vm tests for new verifier.

+++ tools/vm-tests-tf/src/dot/junit/opcodes/invoke_super_range/d/TSuper.d
@@ -31,7 +31,7 @@
     
 .method public toInt(F)I 
 .limit regs 3
-    int-to-float v0, v2
+    float-to-int v0, v2
     return v0
 .end method
